refactor: move RecorderGuard to common
