+++ crates/common/Cargo.toml
@@ -6,10 +6,12 @@ edition = "2021"
 
 [features]
 full-serde = []
+test-utils = ["dep:metrics"]
 
 [dependencies]
 bitvec = "0.20.4"
 hex-literal = "0.3"
+metrics = { version = "0.20.1", optional = true }
 rusqlite = { version = "0.27.0", features = ["bundled"] }
 serde = { version = "1.0.130", features = ["derive"] }
 serde_json = { version = "1.0.68", features = ["arbitrary_precision", "raw_value"] }
